Address review feedback from kdelay
- Use _looks_like() helper to cover both enum.auto() and auto()
  (unqualified form after 'from enum import auto')
- Move new tests from test_brain.py to tests/brain/test_enum.py
  inside EnumBrainTest, alongside existing auto() coverage
- Add test for unqualified auto() spelling
- Clarify comment: value 1 is used for type inference only;
  runtime values for later members will differ
